separate all validations from configure api and create separate modules for each Class
