+++ api/python/provisioner/commands/validator/__init__.py
@@ -25,6 +25,10 @@ from .validator import (FileValidator,
                         ReleaseInfoValidator
                         )
 
+from .network import NetworkParamsValidator
+from .node import NodeParamsValidator
+from .enclosure import StorageEnclosureParamsValidator 
+
 __all__ = [
     "FileValidator",
     "DirValidator",
@@ -34,5 +38,8 @@ __all__ = [
     "YumRepoDataValidator",
     "ContentFileValidator",
     "ReleaseInfoContentScheme",
-    "ReleaseInfoValidator"
+    "ReleaseInfoValidator",
+    "NetworkParamsValidator",
+    "NodeParamsValidator",
+    "StorageEnclosureParamsValidator"
 ]
